squash "timeout: do not handle timeouts while scheduling"
timeout: consider age of local time value

It is normal that the now-time value that is held in the
Alarm_timeout_scheduler is always a little bit behind the actual time of
its time source. But when we schedule an absolute timeout without
considering this fact, the end-time for the timeout, that is calculated
using the local time value, is also smaller than the expected end-time.
This can also lead to directly triggering timeouts that should have
triggered with a certain delay.

As it is not trivial to update the local time value while scheduling a
timeout _without_ calling other timeout handlers, we simply raise the
duration of the new timeout by the age of the local time value.

Issue #2704